合同的机器单机，装置价格等，支持小数点类型。

+++ src/main/java/com/eservice/api/model/contract/Equipment.java
@@ -4,7 +4,8 @@ public class Equipment {
     // 机器装置信息，以JSON字符串的形式进行保存（name/number/price）
     private String name;
     private Integer number;
-    private Integer price;
+//    private Integer price;
+    private Double price;
     private String type;
 
     public void setName(String name) {
@@ -23,11 +24,11 @@ public class Equipment {
         return number;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
